Add confirm to Delete post

+++ admin/includes/view_all_posts.php
@@ -125,7 +125,7 @@ if(isset($_POST['checkBoxArray'])){
               echo "<td>{$post_date}</td>";
               echo "<td> <a href='../post.php?p_id={$post_id}'>View Post</a></td>";
               echo "<td> <a href='?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-              echo "<td> <a href='?delete={$post_id}'>Delete</a></td>";
+              echo "<td> <a onClick=\" javascript: return confirm('Are You Sure To Delete Post'); \" href='?delete={$post_id}'>Delete</a></td>";
               echo "</tr>";
 
             }
